final version without testing

+++ routes/folders.js
@@ -1,16 +1,19 @@
 'use strict';
 
 const express = require('express');
-const router = express.Router();
-
 const mongoose = require('mongoose');
+const passport = require('passport');
 
 const Folder = require('../models/folder');
 const Note = require('../models/note');
 
+const router = express.Router();
+
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/folders', (req, res, next) => {
-  Folder.find()
+  const userId = req.user.id;
+
+  Folder.find({ userId })
     .sort('name')
     .then(results => {
       res.json(results);
@@ -23,6 +26,7 @@ router.get('/folders', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/folders/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -30,7 +34,7 @@ router.get('/folders/:id', (req, res, next) => {
     return next(err);
   }
 
-  Folder.findById(id)
+  Folder.findById({ userId })
     .then(result => {
       if (result) {
         res.json(result);
@@ -46,8 +50,9 @@ router.get('/folders/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/folders', (req, res, next) => {
   const { name } = req.body;
+  const userId = req.user.id;
 
-  const newFolder = { name };
+  const newFolder = { name, userId };
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -73,6 +78,7 @@ router.post('/folders', (req, res, next) => {
 router.put('/folders/:id', (req, res, next) => {
   const { id } = req.params;
   const { name } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -87,7 +93,7 @@ router.put('/folders/:id', (req, res, next) => {
     return next(err);
   }
 
-  const updateFolder = { name };
+  const updateFolder = { name, userId };
 
   Folder.findByIdAndUpdate(id, updateFolder, { new: true })
     .then(result => {
@@ -109,13 +115,14 @@ router.put('/folders/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/folders/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   // Manual "cascading" delete to ensure integrity
-  const folderRemovePromise = Folder.findByIdAndRemove({ _id: id });  // NOTE **underscore** _id
+  const folderRemovePromise = Folder.findByIdAndRemove({ _id: id, userId });  // NOTE **underscore** _id
   // const noteRemovePromise = Note.deleteMany({ folderId: id });
 
   const noteRemovePromise = Note.updateMany(
-    { folderId: id },
+    { folderId: id, userId },
     { '$unset': { 'folderId': '' } }
   );
